fix: split client-safe flag constants out of lib/admin/feature-flags.ts
The client "use client" flags admin page imported KNOWN_OPERATIONAL_FLAG_KEYS
from feature-flags.ts, which also imports the server-only postgres db client.
Bundling that module for the browser pulled in Node builtins (fs/net/perf_hooks)
that Turbopack can't resolve client-side, breaking the production build (and
consequently the E2E job that depends on it). Move the pure constants/validator
into lib/admin/feature-flag-keys.ts (no db import) and re-export from
feature-flags.ts for existing server-side callers.
